storeapi/requirement.txt
asgiref==3.4.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2.4
django-filter==2.4.0
django-versatileimagefield==2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-errors==0.0.1
idna==2.10
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
python-magic==0.4.24
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6